requirements.txt
# Core dependencies for AI Deepfake Project
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
tensorflow>=2.13.0

# Web interface dependencies
Flask>=2.3.0
Werkzeug>=2.3.0

# Data download
kagglehub>=0.1.0
